pyproject.toml
[tool.poetry]
name = "instargram-server-django"
version = "0.1.0"
description = ""
authors = ["NamWook Kim <kimnamwook08@gmail.com>"]
readme = "README.md"
packages = [{include = "instargram_server_django"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.7"
djangorestframework = "^3.14.0"


[tool.poetry.group.dev.dependencies]
autopep8 = "^2.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
